requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
email-validator==2.1.1
Faker==24.9.0
pytest==8.2.0
